# api/swagger.yaml
swagger: "2.0"
info:
  description: "This api server is using to provide data to the tindev app"
  version: "1.0.0"
  title: "Tindev API"
  termsOfService: "http://swagger.io/terms/"
  contact:
    email: "Thibaud.besseau@heig-vd.ch"
  license:
    name: "Apache 2.0"
    url: "http://www.apache.org/licenses/LICENSE-2.0.html"
host: "localhost:5000"
basePath: "/api/"
tags:
- name: "project"
  description: "Everything about your project"
- name: "register"
  description: "Used to add new user"
- name: "apply"
  description: "Apply to a project"
- name: "conversation"
  description: "Conversation"
- name: "profile"
  description: "Everything about your project"
schemes:
- "https"
- "http"
paths:
  /project:
    post:
      tags:
      - "project"
      summary: "Add a new project"
      description: "Propose a new project"
      operationId: "addProject"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "Project object that needs to be added"
        required: true
        schema:
          $ref: "#/definitions/Project"
      responses:
        200:
          description: "OK"
        422:
          description: "Unprocessable Entity"
      security:
      - tindev_auth:
        - "authentificated:user"
    put:
      tags:
      - "project"
      summary: "Update an existing project"
      description: ""
      operationId: "updateProject"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "Project object that needs to be updated"
        required: true
        schema:
          $ref: "#/definitions/Project"
      responses:
        200:
          description: "OK"
        404:
          description: "Project not found"
        422:
          description: "Validation exception"
securityDefinitions:
  tindev_auth:
    type: "oauth2"
    authorizationUrl: "http://localhost:500/auth/login"
    flow: "implicit"
    scopes:
      authentificated:user: "authorized user"
  api_key:
    type: "apiKey"
    name: "api_key"
    in: "header"
definitions:
  Project:
    type: "object"
    properties:
      id:
        type: "integer"
        format: "int64"
      projectName:
        type: "string"
        description: "Project Name"
      amount:
        type: "integer"
        format: "int32"
      description:
        type: "string"
        description: "Project Description"
      timeEstimated:
        type: "string"
        format: "date-time"
      technologies:
        type: "string"
        description: "Technologies used in this project"
      status:
        type: "string"
        description: "Project status"
        enum:
          - "Proposed"
          - "Approved"
          - "Work in progress"
          - "Delivered" 
      deadline:
        type: "string"
        format: "date-time"
    xml:
      name: "Project"
externalDocs:
  description: "Find out more about Swagger"
  url: "http://swagger.io"